Add Client and MySQL DB

+++ src/main/java/com/sbapp/todo/repo/CommonRepository.java
@@ -6,7 +6,7 @@ public interface CommonRepository <T>{
     public T save(T domain);
     public Iterable<T> save(Collection<T> domains);
     public void delete(T domain);
-    public T findById(Integer id);
+    public T findById(Long id);
     public Iterable<T> findAll();
 
 }
