+++ src/main/java/com/sbapp/todo/repo/ToDoJpaRepository.java
@@ -3,5 +3,6 @@ package com.sbapp.todo.repo;
 import com.sbapp.todo.model.ToDo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ToDoJpaRepository extends CrudRepository<ToDo, Integer> {
+public interface ToDoJpaRepository extends CrudRepository<ToDo, Long> {
+
 }
